refactor: move the upload gate into features/upload; share probe readers via utils/probe
The three gate modules (upload-contract, faststart, ensure-upload-contract) are
upload-domain logic consumed only by upload-manager, so they live with it in
features/upload/ alongside tus-client and beat-manifest rather than diluting
utils/. Their tests move with them; relative imports keep the jest doubles
working unchanged.

The probe readers (displaySize, effectiveFps, is10Bit, HDR_TRANSFERS) that
upload-contract borrowed from import-normalization now live in utils/probe.ts:
the upload contract no longer reaches backwards into the import pipeline, and
the two policies read a probe through one shared lens instead of drifting
apart. import-normalization's exports revert to module-private.

No behavior change: tsc clean, expo lint clean, 175/175 tests pass including
the e2e suite.
