+++ src/utils/import-normalization.ts
@@ -1,5 +1,7 @@
 import type { CompressOptions, VideoProbeResult } from 'react-native-video-trim';
 
+import { displaySize, effectiveFps, HDR_TRANSFERS, is10Bit } from './probe';
+
 /**
  * Import normalization policy (§ imports).
  *
@@ -44,35 +46,10 @@ export const NORMALIZE_MAX_BITRATE = 8_000_000;
  * output, uploads) is standardized on H.264 for universal browser playback — HEVC imports
  * are re-encoded once at import time rather than leaking into merged artifacts. */
 const NATIVE_VIDEO_CODECS = new Set(['h264']);
-/** HDR transfer functions: HLG (iPhone camera default) and PQ (HDR10 / Dolby Vision 8.x). */
-export const HDR_TRANSFERS = new Set(['arib-std-b67', 'smpte2084']);
 
 export type ImportDecision =
   | { action: 'passthrough' }
   | { action: 'normalize'; options: Partial<CompressOptions>; reasons: string[] };
-
-/**
- * True for 10-bit pixel formats. FFmpeg names these with a `10`/`10le`/`10be` bit-depth
- * suffix (yuv420p10le, p010le, ...) — matching the suffix rather than a bare `includes('10')`
- * keeps 8-bit chroma-subsampling names like `yuv410p` from being misclassified.
- */
-export function is10Bit(pixelFormat: string): boolean {
-  return /10(le|be)?$/.test(pixelFormat);
-}
-
-/** Effective fps for the decision: average when known (catches VFR), else nominal. */
-export function effectiveFps(probe: VideoProbeResult): number {
-  return probe.averageFps > 0 ? probe.averageFps : probe.nominalFps;
-}
-
-/** Display (post-rotation) dimensions: a 90/270 rotation swaps coded width/height. */
-export function displaySize(probe: VideoProbeResult): { width: number; height: number } {
-  const swapped = probe.rotation % 180 !== 0;
-  return {
-    width: swapped ? probe.height : probe.width,
-    height: swapped ? probe.width : probe.height,
-  };
-}
 
 /**
  * Decide how an imported clip enters the draft: byte-for-byte passthrough, an audio-only
